requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
cloudpickle==2.0.0
cycler==0.11.0
dgl-cu101==0.6.1
dgllife==0.2.9
future==0.18.2
hyperopt==0.2.7
idna==3.3
joblib==1.1.0
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.2
packaging==21.3
pandas==1.1.5
Pillow==8.3.1
py4j==0.10.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
rdkit-pypi==2021.9.4
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8